# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DeployMonster API
  version: 1.4.0
  description: |
    Self-hosted PaaS API for managing applications, deployments, domains,
    secrets, teams, and infrastructure. Single binary, modular monolith
    with event-driven architecture.
  contact:
    name: DeployMonster
    url: https://deploy.monster
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8443
    description: Local development server
  - url: https://{host}
    description: Production server
    variables:
      host:
        default: deploy.monster
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/domains:
    get:
      operationId: listDomains
      tags: [Domains]
      summary: List all custom domains
      responses:
        "200":
          description: Domain list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Domain"
    post:
      operationId: createDomain
      tags: [Domains]
      summary: Add a custom domain to an application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_id, domain]
              properties:
                app_id:
                  type: string
                  format: uuid
                domain:
                  type: string
                  example: app.example.com
                auto_ssl:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Domain added
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Domain"
  /api/v1/domains/{id}:
    delete:
      operationId: deleteDomain
      tags: [Domains]
      summary: Remove a custom domain
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: Domain removed
...
  /api/v1/secrets:
    get:
      operationId: listSecrets
      tags: [Secrets]
      summary: List all secrets (values masked)
      responses:
        "200":
          description: Secret list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                        updated_at:
                          type: string
                          format: date-time
    post:
      operationId: createSecret
      tags: [Secrets]
      summary: Create or update a secret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, value]
              properties:
                name:
                  type: string
                  pattern: "^[A-Z_][A-Z0-9_]*$"
                value:
                  type: string
      responses:
        "201":
          description: Secret created
  /api/v1/secrets/{id}:
    delete:
      operationId: deleteSecret
      tags: [Secrets]
      summary: Delete a secret
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Secret deleted

  # ── Billing ───────────────────────────────────────────
...
  /api/v1/admin/api-keys:
    get:
      operationId: listApiKeys
      tags: [Admin]
      summary: List API keys
      responses:
        "200":
          description: API key list
    post:
      operationId: createApiKey
      tags: [Admin]
      summary: Generate a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                expires_in_days:
                  type: integer
      responses:
        "201":
          description: API key created (shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    description: Full API key (only shown once)
                  prefix:
                    type: string
                  name:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
  /api/v1/admin/api-keys/{prefix}:
    delete:
      operationId: deleteApiKey
      tags: [Admin]
      summary: Revoke an API key by prefix
      parameters:
        - name: prefix
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: API key revoked
...
  /api/v1/projects/{id}:
    get:
      operationId: getProject
      tags: [Apps]
      summary: Get project details
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Project details
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProject
      tags: [Apps]
      summary: Delete a project
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: Project deleted

  # ── MCP ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key generated from admin panel
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource identifier
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: not found
  schemas:
...
    Domain:
      type: object
      properties:
        id:
          type: string
          format: uuid
        app_id:
          type: string
          format: uuid
        domain:
          type: string
          example: app.example.com
        ssl_enabled:
          type: boolean
        ssl_expires_at:
          type: string
          format: date-time
        verified:
          type: boolean
        created_at:
          type: string
          format: date-time
...